Continue on seeing '@' in lexer advance.
Summary:
When the SourceErrorManager is using the SimpleDiagHandler,
which allows at most 1 error to be reported, and ignores following errors,
when we see the '@' in the template literal while lexing here,
we error and forceEOF (we've hit the error limit).
We have to then `continue` on encountering an error in the loop
to set the TokenKind to eof.
Otherwise, the lexer gets in an inconsistent state.

Reviewed By: tmikov

Differential Revision: D23409028

fbshipit-source-id: 0ba17c644926326198d5c150a19ebb92e15a1499