Remove OG bump alloc mode
Summary:
Since we only check for bump alloc after the free list is completely
exhausted, it would be simpler to just add the segment to the free
list when it is added to OG. This way, we will see the block as we
scan the free list and split it as needed.

This lets us eliminate a few special cases:
1. Calculating allocated bytes.
2. Checking bump alloc mode when sweeping.
3. Scanning segments for bump alloc opportunities every time we can't
find something in the free list.

Reviewed By: dulinriley

Differential Revision: D23390878

fbshipit-source-id: a3d9bb21f34716729103eb2579039e335ce6a3a2