Add write barriers to all places where an array length is shortened
Summary:
In Hades, a write barrier is needed when a pointer becomes unreachable. This
primarily happens when the GC has a different set of things that might be marked.
Currently, the only way that can happen is if an array's length is reduced.

Whenever that can occur, call a new API on GCHermesValue: `unreachableWriteBarrier`,
to alert the GC of the change.

Reviewed By: davedets

Differential Revision: D23362654

fbshipit-source-id: 4125f899774336e60aafd7282310aa7d25388a2b