Add heap segment extents to CrashManager custom data
Summary:
For each segment added to Hades, update some CrashManager metadata
saying where the start of each segment is. End addresses can be computed
because segments are a fixed size. This can be used in crash data to see if
a pointer is into the JS heap.

Give YG a special name, because in a crash situation it's helpful to know
whether something is a young object or an old object.

Reviewed By: neildhar

Differential Revision: D23514115

fbshipit-source-id: 5e27eae0eb8e24130e1ad85f35a48256e0673ffa