Lazily allocate additional segments for GC
Summary:
Add an interface to count the number of segments that we are willing
to lazily allocate in the old gen. The interface tries to resemble
std containers (like other parts of `OldGen`).

The effective capacity of the container is exposed via the `capacity`
method, which is currently only used by YG collections to determine
if it is appropriate to start an OG collection.

Reviewed By: dulinriley

Differential Revision: D23470930

fbshipit-source-id: abdd9ccec25ce39b1fdab7dd7e505f05654b1985