Add ExponentialMovingAverage and track survival ratios in Hades
Summary:
It will be useful in later changes to have an idea of how many bytes are
surviving collections in the young gen and the old gen of Hades.
This can influence, for example, the speed of marking in incremental mode.

Add an `ExponentialMovingAverage` class that represents an exponentially weighted moving
average, where `weight` is the factor applied to new data. Use this for the survival
ratios of both YG and OG.

Currently these stats are unused, but they will be used later.

Reviewed By: neildhar

Differential Revision: D23604717

fbshipit-source-id: 96c00eede4e3275d504ca56547eddb75b48ee13d