Have Hades pick an appropriate drain size based on YG survival rates
Summary:
Before, Hades always used a drain "bucket" size of 128 cells.
This didn't adapt very well to changing circumstances, such as if
there was a lot of allocations that survive into OG. The more things
that survive into OG, the more time we should spend marking and
freeing cells so there will be space.

Switch to using a bytes-based measurement, and calculate a suitable
drain rate based on various factors:
* Expected amount of bytes to mark in the OG
* Number of YG collections until the OG is full

Cap the drain rate at a minimum of 8 KB to ensure even if YG collections
are promoting next to nothing, that the collection will still eventually complete.

Reviewed By: davedets

Differential Revision: D23295898

fbshipit-source-id: d7144b0bf7833dd3fd03a609b4a6a490c630c753