Fix Array.from using defineOwnProperty instead of defineOwnComputedPrimitive
Summary:
`Array.from` was using `JSObject::defineOwnProperty` to create properties on the new array.
Using that API creates a Symbol and a StringPrimitive for
a simple number. It also triggers later slow-paths in the VM, because the "fast
index" flags aren't set correctly.

Fix this by changing to `JSObject::defineOwnComputedPrimitive`.

Reviewed By: tmikov

Differential Revision: D23770063

fbshipit-source-id: 1f0412be482793275408eb21ccbcb2f869ca65bb