Have to_node use the ID instead of the index to avoid confusion
Summary:
I had confused myself reading this output, thinking "that to_node seems
to be pointing nowhere!" Turns out it was an *index*, not the *ID* of the node
pointed to.

Fix this so I won't get confused again. It's a lot easier to grep through a json
file for `"id": 1234` than to have to pop into node.js to parse the JSON file.

Reviewed By: neildhar

Differential Revision: D23879549

fbshipit-source-id: d6c53e90bc779bff41a3a081b3093ee4acbe6892