Simplify usage of BasedPointer
Summary:
Move all of the compressed pointer ifdefs into the PointerBase code.
This allows us to always write code as though the compressed pointers
are enabled, and rely on the different internal implementations to
turn compressed pointers on/off.

This will also make it easier to use BasedPointers directly,
specifically in GC code, where using a GCPointer doesn't make sense.

Reviewed By: dulinriley

Differential Revision: D23674752

fbshipit-source-id: 50f41d3708e08011a0b32aaf0c865a46285f02a9