Remove astral character assert in U16MatchCharICase32Insn
Summary:
We currently have an assert that characters encountered in a
`U16MatchCharICase32Insn` must be astral. This is not necessarily the
case since unpaired low/high surrogates may also be encoded in
`U16MatchCharICase32Insn`instructions. We also don't have such an
assert for the case sensitive variant `U16MatchChar32Insn`, which is
generated in almost the same way.

Reviewed By: avp

Differential Revision: D23927543

fbshipit-source-id: 01593e2c434676be22ee4333740b8c448b7276cf