Fix issues when getting timezone string
Summary:
Always set the time_t variable so it is valid. This is needed because
MSVC will set localtime to null if the variable is invalid.

In addition, while looking back at this code, I realised that
localtime is writing to some internal shared buffer. Access to this
buffer is not guaranteed to be thread safe. We should create our own
buffer and use localtime_r/localtime_s instead.

Reviewed By: mhorowitz

Differential Revision: D23919559

fbshipit-source-id: 8b679936c4c9467f19eb5f56752972c85601af95