Avoid an assert when defining a prop on a Proxy array target
Summary:
JSProxy::defineOwnProperty was implemented to call
JSObject::defineOwnProperty.  Recently, this would start to assert if
a prop which looks like an array index was used.  In order to avoid
this, change JSProxy::defineOwnProperty to take a Handle<> as the
name, and call JSObject::defineOwnComputedProperty, which knows how to
handle index-ish properties.  This has the downside of converting a
named property from SymbolID to Handle to SymbolID in some proxy
cases, but the perf loss is preferable to writing a SymbolID overload
of JSProxy::defineOwnProperty.

closes #371

Reviewed By: tmikov

Differential Revision: D23914310

fbshipit-source-id: 633ad089bccf1aef559ffb95e05deaf6a22c52af