Generate Hermes AST visitor keys from AST definitions
Summary:
We will need to traverse the Hermes AST in order to convert it to a proper ESTree or Babel AST. Let's autogenerate the information needed to traverse the Hermes AST directly from the Hermes AST definitions similar to how this is done for the JS builder methods. We can use the C preprocessor to generate a map of the visitor keys in JS, mapping from the node's type to an object containing all of its node and node list children (string, boolean, and numeric fields are ignored). This allows us efficiently traverse the AST by looking up the visitor keys for each node when it is visited, and subsequently visiting the node child or all of the node list children.

Added autogeneration of these visitor keys to the `build.sh` script, being sure to format and then sign the generated file.

Also included a simple implementation of traversing the AST using these visitor keys in `HermesASTVisitor`. This is not yet used, but will be a foundation for the handful of conversions we need to do to modify the AST to be Babel or ESTree compliant.

Reviewed By: avp

Differential Revision: D23768640

fbshipit-source-id: 6ec34b7505075d3b3298c13f27f385c97e766d15