Set up adapters to ESTree and Babel and fix source location format
Summary:
The Hermes parser compiled to WASM currently outputs a "Hermes AST", which is very close to (though not quite equivalent when transferred to JS) to an ESTree AST. We want to be outputting a proper ESTree AST or a Babel AST in JS so that this parser can replace usages of the Babel parser as well as be used in ESLint. Luckily there are not many differences here and the few that exist are minor, but we must be sure to handle them properly.

Let's set up the basic adapters to transform the Hermes AST to a compliant ESTree or Babel AST. This involves walking the Hermes AST in JS using the Hermes AST's visitor keys. We can then update the AST in place for efficiency. A user can specify whether they want a Babel AST output by specifying `{babel: true}` in the optional second argument to the top-level parse API. If `babel: true` is not specified, an ESTree AST is returned by default.

The first modification we must make is actually an artifact of the transfer of the AST from JS to WASM. When being transferred to JS from WASM we put all source location information in the same object which is constructed with the following structure, where `rangeStart` and `rangeEnd` are the offsets into the source text:

```
{
  start: { line, column },
  end: { line, column },
  rangeStart,
  rangeEnd,
}
```

Babel places the start and end range offsets directly on the node itself under the `start` and `end` keys.

The ESTree spec does not require that nodes include the start/end range, however let's add a `range: [rangeStart, rangeEnd]` property as this will be required for using the ESTree AST in ESLint [docs](https://eslint.org/docs/developer-guide/working-with-custom-parsers).

This also involved updating the column numbers to be 0-indexed (they were previously being constructed as 1-indexed).

Reviewed By: gkz

Differential Revision: D23796135

fbshipit-source-id: a13451207ae5a575ffc3ff6f0e64614246589c4c