Convert directives to Babel AST format
Summary:
The Hermes AST currently stores directives in the AST following the ESTree spec, where directives are a special type of `ExpressionStatement` where the expression is a string literal and the `directive` property is set ([spec](https://github.com/estree/estree/blob/master/es5.md#directive)). However we want to be able to also output directives in the Babel AST format, where directives are stored as `Directive` nodes containing a `DirectiveLiteral` in a separate array on `Program` and `BlockStatement` nodes ([spec](https://github.com/babel/babel/blob/master/packages/babel-parser/ast/spec.md#directive)).

This involved adding a step to the `HermesToBabelAdapter` which converts directives from the ESTree to Babel format. This step only needs to run on `Program` and `BlockStatement` nodes, where it removes directive nodes from the `body` property and adds them in the new format under the `directives` property. We can determine if an `ExpressionStatement` is actually a directive by checking whether its `directive` property is set, and we can fail fast when we find an `ExpressionStatement` without a directive as directives must precede all expression statements.

Reviewed By: gkz

Differential Revision: D23826092

fbshipit-source-id: a4b2b62890f4faba3bb15a1b24e4b980a58621b4