Add boolean method field to Property node
Summary:
`Property` nodes in the AST are currently missing two fields defined in the ESTree spec (ES6 extension: https://github.com/estree/estree/blob/master/es2015.md#objectpattern). One of these is the method field, which is a boolean field that denotes whether or not this field was defined using method syntax. Note that only properties defined like `{ foo() {} }` are considered methods - getters and setters are not considered methods for the purpose of this field (they are noted by the `kind = get` or `kind = set` field).

This field is needed to output a full, compliant ESTree AST from the parser, such as in the Hermes parser compiled to WASM. In addition, we need this field in the Hermes parser compiled to WASM so that we can convert ESTree's Property nodes to Babel's `ObjectProperty` and `ObjectMethod` nodes, which need to know whether a particular property is defined as a method.

We can simply set this field while parsing, depending on how the property was defined.

Reviewed By: avp

Differential Revision: D23844482

fbshipit-source-id: f339c39fbc0c0e41cbe160db32bdb1d4b72ef63f