Add option to back GC segments with huge pages
Summary:
Add a flag that allows the VM to use huge pages for GC segments. This
comes at the cost of an increased memory footprint since we no longer
return memory to the VM at a page granularity. However, this may
improve performance in certain cases because:
1. Larger pages imply fewer TLB entries to cover the same addressable
region. This reduces TLB misses.
2. More of the space we need is reserved upfront, reducing minor page
faults.

Reviewed By: dulinriley

Differential Revision: D23669692

fbshipit-source-id: adf1c25beae38aab49ebc6f48ad714e8b3c5aa4e