Replace cjsModuleOffset with segmentID
Summary:
Introduces the concept of a "segment ID" to the runtime and bytecode.

Previously, each BytecodeModule / RuntimeModule had a `cjsModuleOffset` which did double duty as (1) the offset of the first CJS module for registration in the Domain, and (2) a unique ID for each BytecodeModule / RuntimeModule for the purpose of stack traces and source maps.

Here we:

* Remove responsibility (1) because the changes to Domain in D23938175 make it unnecessary.
* Use the input segment ID (from the CommonJS metadata) as the segment ID in the corresponding BytecodeModule / RuntimeModule.
* Rename the resulting concept to `segmentID`.

Reviewed By: avp

Differential Revision: D23940758

fbshipit-source-id: 7f70179a42ca7aceb7c6a5dcfe2ad12141ddfb74