Add sourceType property to Program node
Summary:
Both the Babel and ESTree ASTs contain a `sourceType` property on the `Program` node which can be either `"module"` or `"script"` depending on whether the source was parsed as an ES6 module.

We can gather the information needed to set this field from within te `HermesASTAdapter`. If we encounter a value import or export, then we know we are in a module, otherwise we are in a script. This is the same logic used by the Babel parser (though in addition the Babel parser checks for top level awaits and `import.meta`, both of which we do not support).

We can simply override the mappers for all import and export statements, and then set the internal `sourceType` property when a value import or export is encountered. We then add this property during `mapProgram` (which required implementing `mapProgram` for the ESTree adapter).

Reviewed By: gkz

Differential Revision: D24183583

fbshipit-source-id: 8b502d0327d4502c125a862c2c987ecf07fdcc99