Remove clang-tidy readability-else-after-return check
Summary:
This check can be a bit too draconian, for example:
```
if (foo) {
  return "foo";
} else {
  return "bar";
}
```

The `readability-else-after-return` check will try to fix this to:
```
if (foo) {
  return "foo";
}
return "bar";
```

This can end up being less readable in some cases, especially for short
else statements. Remove enforcement of this from Hermes's clang-tidy.

Perhaps in the future, if this check allows configuring the minimum
number of lines before triggering the warning, we can re-enable it.

Reviewed By: neildhar

Differential Revision: D24209615

fbshipit-source-id: fe8de3b043ef32d83743db5707b5911d0b1191c3