Bump clang-tidy readability-braces-around-statements to 3 lines
Summary:
The `readability-braces-around-statements` forces curly braces around
statements after a condition. It's configurable to have a threshold of
when it applies.

Previously, the threshold was 2 lines, which allowed code like this:
```
if (foo)
  return;
```

Bump to 3 lines to allow code like this:
```
if (foo)
  return 1;
else
  return 2;
```

The documentation says this about `ShortStatementLines`:
> The number of lines is counted from the end of condition or
initial keyword (`do`/`else`) until the last line of the inner statement

Differential Revision: D24210313

fbshipit-source-id: 7590d55b6c6ed192fc27ceb71b5188870fca1718