Allow duplicate modules across segments
Summary:
Makes it possible for a single CJS module to be included in multiple segments (and therefore multiple resulting HBC files). For example, we can have the following segment layout:

```
0: (main segment)
- main.js
1:
- foo.js
- bar.js
2:
- foo.js
- baz.js
```

When any file in either segment 1 or 2 does `require('foo.js')` the correct CJS module is loaded. The first call to `require('foo.js')` will execute `foo.js` and store the exports.

Building on the Domain, RuntimeModule, BytecodeModule, and IR generation changes in the lower diffs, this diff enables CompilerDriver to handle this scenario by explicitly checking for duplicate module paths and assigning module IDs accordingly.

Reviewed By: avp

Differential Revision: D23944854

fbshipit-source-id: 59e260a8bd35f532dbac9ef1b27850ef1cad87b7