Add an OOM to the Hades alloc path for a large size
Summary:
There was an assert in `HadesGC::allocWork` that might not always be true,
and we'd like to know when it isn't as it indicates a bug in the VM.

Specifically, if any one allocation wants to be bigger than a heap segment,
it'll never successfully allocate. Add a check for this after the first YG collection
fails, as it's guaranteed to reach that far for a true super-segment allocation, and
it won't penalize the common path.

Add a unit test to GCOOMNCTest, since that test file is now no longer NCGen
specific, remove the "NC" part. Also fix up the naming convention in the test
file to match the rest of Hermes.

Reviewed By: neildhar

Differential Revision: D24261877

fbshipit-source-id: 1a489585b71d49f743de3653a79256d932dba1e0