Enum body locations include explicit type
Summary:
Let's update the source location for some nodes that differ from the source locations reported by Flow and Babel. One instance of this is the source location for enum bodies that contain explicit types (e.g. `enum E of string {}`). Flow and Babel consider the explicit type (`of string`) as part of the enum body's source location, whereas Hermes excludes the explicit type from the enum body's source location.

Let's update Hermes to match the behavior of Flow and Babel here. We can accomplish this by refactoring the `explicitType` parameter to `parseEnumBody` from a bool signaling whether an explicit type was provided, to an optional source location for the start of the explicit type if one exists. Then we can check if a location was passed to determine both whether there was an explicit type for this enum body, and if so at what location it started.

Reviewed By: avp

Differential Revision: D24151986

fbshipit-source-id: 60e838708420a8539b789abe7121b53a051f19de