Create GenGCHeapSegment
Summary:
There is currently a lot of GenGC specific functionality embedded in
`AlignedHeapSegment`. This adds a lot of redundant code when compiling
with Hades and forces us to use private inheritance in Hades'
`HeapSegment` to hide that functionality.

In addition, it creates circular dependences since a lot of these
features like marking and compacting are fairly heavy, and therefore
likely rely on other features that then in turn rely on
`AlignedHeapSegment`. This change should break some of those circular
dependences.

This diff just does the bare minimum to rename all uses of
`AlignedHeapSegment` to `GenGCHeapSegment`. A subsequent change will port
the actual functionality over.

Reviewed By: dulinriley

Differential Revision: D24382680

fbshipit-source-id: 76e27e58edc190e6ae4e2950ccaff9a79faede64