Add total allocated bytes to Hades
Summary:
Total allocated bytes is a measure of all the allocations a program does
over its lifetime. This metric is similar to a sum of all calls to `malloc`, even
for memory that is `free`'d.
It can be used to see if a VM optimization results in fewer allocations.

To avoid adding cost to each individual allocation, track in bulk at each YG
emptying. Add `youngGen().used()` whenever the value is queried to avoid
undercounting allocations that haven't been collected yet.

Reviewed By: neildhar

Differential Revision: D24426063

fbshipit-source-id: 336c93bbfca9e2b00fa5c2da1309af8b45376912