Fix tripwire test for incremental Hades
Summary:
The tripwire test currently does not trigger the tripwire for
incremental collections. This happens because we currently only check
the stats for a completed incremental OG collection during the next
YG collection.

Note that adding the tripwire check to the non-concurrent GC branch
in `waitForCollectionToFinish` is not sufficient because the OG
collection may complete immediately if there are very few objects,
and therefore it won't be active once `waitForCollectionToFinish` is
called.

This could technically happen with an OG collection too, if an OG
collection completes fully before a call to
`waitForCollectionToFinish`, then the tripwire check will happen at
the next YG instead. However, for the purposes of this test, that is
not possible, because the OG has to synchronise with the mutator in
`waitForCollectionToFinish` before it can complete the collection.

Reviewed By: dulinriley

Differential Revision: D24411327

fbshipit-source-id: 8cbe6e535829d9485b5947d46ed9846c9e0988e3