Handle ASI in class properties.
Summary:
Class properties result in some complex semicolon insertion hazards.
```
class C {
  async
}
```
must insert a `;` after the `async` because otherwise the parse would
fail.

Relatedly:
```
class C {
  async
  foo
}
```
must be parsed as two separate properties, because there is a newline
between `async` and `foo`, but this is handled differently for `get`
and `set`, which have no [no newline here] requirement in the grammar.

Handle this by specially checking for the next token and manually
applying the ASI rules when parsing class properties.

Reviewed By: tmikov

Differential Revision: D23084521

fbshipit-source-id: 1b0586c7e9393d6507a441809a57ecca9f56a37a