Store hashbang comments with unique comment type
Summary:
Currently, the Hermes parser stores hashbang comments (comments that begin with `#!` and start at the first byte of the file) as regular line comments. Let's introduce a new `Hashbang` comment type for these comments as they are not actually regular line comments, and so that they can be easily distinguished.

This will be used to distinguish these hashbang comments from other line comments in the Hermes parser compiled to WASM.

Reviewed By: avp

Differential Revision: D24352504

fbshipit-source-id: 653a8c2efd7bd188ee89a64c465905ad2ccdd6ca